requirements.txt
Flask
requests
beautifulsoup4
newspaper3k
praw
